+++ src/main/java/com/tsystems/javaschool/ecare/dao/IAbstractDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface IAbstractDAO<T> {
     public T saveOrUpdate(T t);
-    public T load(long id);
+    public T load(int id);
     public void delete(T t);
     public List<T> getAll();
     public void deleteAll();
